pyproject.toml
[project]
name = "my-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "“leejiyun1”",email = "“poi20701556@gmail.com”"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-decouple (>=3.8,<4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
